temp, not working

+++ src/util.c
@@ -1,6 +1,6 @@
 #include "util.h"
 
-Vector2f createVector(float x, float y) {
+Vector2f createVector(int x, int y) {
     Vector2f temp;
     temp.x = x;
     temp.y = y;
